fdt: Pass the device revision number through devicetree
As we did for the serial number, also pass the revision number:

Before device-tree, the device revision number used to be passed to the
kernel using ATAGs (on ARM). This is now deprecated and all the handover
to the kernel should now be done using device-tree. Thus, this passes the
revision-number property to the kernel using the serial-number property of
the root node, as expected by the kernel.

Unlike the the serial number is the revision number is not a string but a
32 bits number.

Signed-off-by: Gregory CLEMENT <gregory.clement@bootlin.com>